Update masking using geoR.R
Uses R packages sp, gstat, raster to make spatial data objects from a data frame and perform inverse distance weighted interpolation on a georeferenced variable. The interpolated grid is masked using geoR::locations.inside() and plotted on an OpensStreetMap basemap using fields::plot.image, and this map is all the output of the code. The lctools package is used to calculate Global and Local Moran's I spatial autocorrelations, with the significant (p<0.05) local Moran's I categorised into high-high, low-low, high-low, and low-high associations and plotted on the map, with a legend which also contains the value of Global Moran's I and the common bandwidth for both types of spatial autocorrelation.